+++ package.json
@@ -10,9 +10,9 @@
   },
   "dependencies": {
     "axios": "^1.7.7",
-    "next": "13.5.6",
-    "react": "^18",
-    "react-dom": "^18"
+    "next": "^13.5.6",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "@types/node": "^20",
